+++ models/transformations/base/t_cableline.sql
@@ -9,7 +9,7 @@
 }}
 
 SELECT
-    row_number() OVER (ORDER BY cl_code) AS id,
+    row_number() OVER (ORDER BY cl_code)::int4 AS id,
     CASE WHEN pg_input_is_valid(NULLIF(cl_cb_code::text, ''), 'varchar(254)') THEN cl_cb_code::VARCHAR(254) ELSE NULL END AS cl_cb_code,
     CASE WHEN pg_input_is_valid(NULLIF(cl_code::text, ''), 'varchar(254)') THEN cl_code::VARCHAR(254) ELSE NULL END AS cl_code,
     geom AS geom