+++ models/transformations/base/t_ebp.sql
@@ -15,7 +15,7 @@
 }}
 
 SELECT
-    row_number() OVER (ORDER BY bp_code) AS id,
+    row_number() OVER (ORDER BY bp_code)::int4 AS id,
     CASE WHEN pg_input_is_valid(NULLIF(bp_abandon::text, ''), 'varchar(1)') THEN bp_abandon::VARCHAR(1) ELSE NULL END AS bp_abandon,
     CASE WHEN pg_input_is_valid(NULLIF(bp_avct::text, ''), 'varchar(1)') THEN bp_avct::VARCHAR(1) ELSE NULL END AS bp_avct,
     CASE WHEN pg_input_is_valid(NULLIF(bp_code::text, ''), 'varchar(254)') THEN bp_code::VARCHAR(254) ELSE NULL END AS bp_code,